+++ database/migrations/2024_03_03_062618_create_receiving_reports.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('receiving_reports', function (Blueprint $table) {
             $table->increments('receiving_report_id');
-            $table->float('total_price');
+            $table->decimal('total_price', 10, 2);
             $table->unsignedInteger('supplier_id');
             $table->unsignedInteger('created_by');
             $table->timestamps();
